lara/cheat_keys: merge cheat keys modules
This merges Lara's original cheats into TRX.
