+++ src/libtrx/include/libtrx/game/lara.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "lara/cheat.h"
+#include "lara/cheat_keys.h"
 #include "lara/common.h"
 #include "lara/const.h"
 #include "lara/enum.h"
