+++ src/libtrx/meson.build
@@ -166,6 +166,7 @@ sources = [
   'game/items/col.c',
   'game/items/common.c',
   'game/items/utils.c',
+  'game/lara/cheat_keys.c',
   'game/lara/common.c',
   'game/lara/hair.c',
   'game/lara/misc.c',
