+++ src/tr1/game/lara/cheat.c
@@ -22,108 +22,6 @@
 
 #include <stdint.h>
 
-void Lara_Cheat_Control(void)
-{
-    static int32_t cheat_mode = 0;
-    static int16_t cheat_angle = 0;
-    static int32_t cheat_turn = 0;
-
-    if (Game_GetCurrentLevel() == GF_GetGymLevel()) {
-        return;
-    }
-
-    LARA_STATE as = g_LaraItem->current_anim_state;
-    switch (cheat_mode) {
-    case 0:
-        if (as == LS_WALK) {
-            cheat_mode = 1;
-        }
-        break;
-
-    case 1:
-        if (as != LS_WALK) {
-            cheat_mode = as == LS_STOP ? 2 : 0;
-        }
-        break;
-
-    case 2:
-        if (as != LS_STOP) {
-            cheat_mode = as == LS_BACK ? 3 : 0;
-        }
-        break;
-
-    case 3:
-        if (as != LS_BACK) {
-            cheat_mode = as == LS_STOP ? 4 : 0;
-        }
-        break;
-
-    case 4:
-        if (as != LS_STOP) {
-            cheat_angle = g_LaraItem->rot.y;
-        }
-        cheat_turn = 0;
-        if (as == LS_TURN_LEFT) {
-            cheat_mode = 5;
-        } else if (as == LS_TURN_RIGHT) {
-            cheat_mode = 6;
-        } else {
-            cheat_mode = 0;
-        }
-        break;
-
-    case 5:
-        if (as == LS_TURN_LEFT || as == LS_FAST_TURN) {
-            cheat_turn += (int16_t)(g_LaraItem->rot.y - cheat_angle);
-            cheat_angle = g_LaraItem->rot.y;
-        } else {
-            cheat_mode = cheat_turn < -94208 ? 7 : 0;
-        }
-        break;
-
-    case 6:
-        if (as == LS_TURN_RIGHT || as == LS_FAST_TURN) {
-            cheat_turn += (int16_t)(g_LaraItem->rot.y - cheat_angle);
-            cheat_angle = g_LaraItem->rot.y;
-        } else {
-            cheat_mode = cheat_turn > 94208 ? 7 : 0;
-        }
-        break;
-
-    case 7:
-        if (as != LS_STOP) {
-            cheat_mode = as == LS_COMPRESS ? 8 : 0;
-        }
-        break;
-
-    case 8:
-        if (g_LaraItem->fall_speed > 0) {
-            if (as == LS_JUMP_FORWARD) {
-                Game_SetIsLevelComplete(true);
-            } else if (as == LS_JUMP_BACK) {
-                Inv_AddItem(O_SHOTGUN_ITEM);
-                Inv_AddItem(O_MAGNUM_ITEM);
-                Inv_AddItem(O_UZI_ITEM);
-                g_Lara.shotgun_ammo.ammo = 500;
-                g_Lara.magnum_ammo.ammo = 500;
-                g_Lara.uzi_ammo.ammo = 5000;
-                Sound_Effect(SFX_LARA_HOLSTER, nullptr, SPM_ALWAYS);
-            } else if (as == LS_SWAN_DIVE) {
-                Item_Explode(g_Lara.item_num, -1, 1);
-                Sound_Effect(SFX_EXPLOSION_CHEAT, &g_LaraItem->pos, SPM_NORMAL);
-                g_LaraItem->hit_points = 0;
-                g_LaraItem->flags |= IF_INVISIBLE;
-            }
-            cheat_mode = 0;
-        }
-        break;
-
-    default:
-        cheat_mode = 0;
-        break;
-    }
-}
-
 void Lara_Cheat_EndLevel(void)
 {
     Game_SetIsLevelComplete(true);
