+++ src/tr2/meson.build
@@ -117,7 +117,6 @@ sources = [
   'game/inventory_ring/vars.c',
   'game/item_actions.c',
   'game/lara/cheat.c',
-  'game/lara/cheat_keys.c',
   'game/lara/col.c',
   'game/lara/common.c',
   'game/lara/control.c',
